move isVideoTypeSupported to utils and update imports
