+++ front-end/src/components/EditorContainer.jsx
@@ -8,7 +8,7 @@ import AddText from './AddText'
 import ColorFilters from './ColorFilters'
 import GifEditor from './GifEditor'
 import useMediaSelection from '../hooks/useMediaSelection'
-import { isVideoTypeSupported } from './videoSupport'
+import { isVideoTypeSupported } from '../utils/videoSupport'
 import CameraCapture from './CameraCapture'
 import PhotoPreview from './PhotoPreview'
 import { exportImageFromBackend } from '../services/backendImageService'
